+++ aes.go
@@ -10,7 +10,6 @@ import (
 	"io"
 
 	"go.innotegrity.dev/errorx"
-	"go.innotegrity.dev/slogx"
 )
 
 // DecryptString decrypts the given block of ciphertext that was encrypted using the EncryptString() function.
@@ -20,15 +19,10 @@ import (
 // The following errors are returned by this function:
 // DecryptionError
 func DecryptString(ctx context.Context, ciphertext, key string) (string, errorx.Error) {
-	logger := slogx.ActiveLoggerFromContext(ctx)
-	errAttr := slogx.ErrorAttrNameFromContext(ctx)
-
 	// decode the Base64-encoded string
 	data, err := base64.StdEncoding.DecodeString(ciphertext)
 	if err != nil {
-		e := NewDecryptionError("failed to decode ciphertext string", err)
-		logger.Error(e.Msg(), slogx.ErrX(errAttr, e))
-		return "", e
+		return "", NewDecryptionErrorWithContext(ctx, "failed to decode ciphertext string", err)
 	}
 
 	// extract the key, nonce and ciphertext
@@ -66,25 +60,19 @@ func DecryptString(ctx context.Context, ciphertext, key string) (string, errorx.
 	// create a new cipher block from the key
 	block, err := aes.NewCipher(sha[0:32])
 	if err != nil {
-		e := NewDecryptionError("failed to generate new cipher", err)
-		logger.Error(e.Msg(), slogx.ErrX(errAttr, e))
-		return "", e
+		return "", NewDecryptionErrorWithContext(ctx, "failed to generate new cipher", err)
 	}
 
 	// create a new GCM
 	aesGCM, err := cipher.NewGCM(block)
 	if err != nil {
-		e := NewDecryptionError("failed to generate GCM", err)
-		logger.Error(e.Msg(), slogx.ErrX(errAttr, e))
-		return "", e
+		return "", NewDecryptionErrorWithContext(ctx, "failed to generate GCM", err)
 	}
 
 	// decrypt the data
 	plaintext, err := aesGCM.Open(nil, nonce, data, cipherKey)
 	if err != nil {
-		e := NewDecryptionError("failed to decrypt the data", err)
-		logger.Error(e.Msg(), slogx.ErrX(errAttr, e))
-		return "", e
+		return "", NewDecryptionErrorWithContext(ctx, "failed to decrypt the data", err)
 	}
 	return string(plaintext), nil
 }
@@ -96,17 +84,12 @@ func DecryptString(ctx context.Context, ciphertext, key string) (string, errorx.
 // The following errors are returned by this function:
 // EncryptionError
 func EncryptString(ctx context.Context, plaintext, key string) (string, errorx.Error) {
-	logger := slogx.ActiveLoggerFromContext(ctx)
-	errAttr := slogx.ErrorAttrNameFromContext(ctx)
-
 	// generate a random key if needed
 	cipherKey := []byte(key)
 	if key == "" {
 		cipherKey = make([]byte, 32)
 		if _, err := rand.Read(cipherKey); err != nil {
-			e := NewEncryptionError("failed to generate random key", err)
-			logger.Error(e.Msg(), slogx.ErrX(errAttr, e))
-			return "", e
+			return "", NewEncryptionErrorWithContext(ctx, "failed to generate random key", err)
 		}
 	}
 
@@ -116,25 +99,19 @@ func EncryptString(ctx context.Context, plaintext, key string) (string, errorx.E
 	// create a new cipher block from the key
 	block, err := aes.NewCipher(sha[0:32])
 	if err != nil {
-		e := NewEncryptionError("failed to generate a new cipher", err)
-		logger.Error(e.Msg(), slogx.ErrX(errAttr, e))
-		return "", e
+		return "", NewEncryptionErrorWithContext(ctx, "failed to generate a new cipher", err)
 	}
 
 	// create a new GCM
 	aesGCM, err := cipher.NewGCM(block)
 	if err != nil {
-		e := NewEncryptionError("failed to generate GCM", err)
-		logger.Error(e.Msg(), slogx.ErrX(errAttr, e))
-		return "", e
+		return "", NewEncryptionErrorWithContext(ctx, "failed to generate GCM", err)
 	}
 
 	// create a nonce from GCM
 	nonce := make([]byte, aesGCM.NonceSize())
 	if _, err = io.ReadFull(rand.Reader, nonce); err != nil {
-		e := NewEncryptionError("failed to generate nonce from GCM", err)
-		logger.Error(e.Msg(), slogx.ErrX(errAttr, e))
-		return "", e
+		return "", NewEncryptionErrorWithContext(ctx, "failed to generate nonce from GCM", err)
 	}
 
 	// encrypt the data and hide the key inside the ciphertext