+++ doc.go
@@ -1,2 +1,5 @@
 // Package cryptox contains functionality for dealing with X509 certificates and cryptography.
+//
+// This package utilizes the go.innotegrity.dev/errorx package to return errors which can contain caller information
+// that can be used with whatever logging platform you prefer to use.
 package cryptox