+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 No unreleased changes
 
+## v0.1.2 (Released 2024-04-01)
+
+* Removed message logging and dependency on `go.innotegrity.dev/slogx`
+* Updated to use new `go.innotegrity.dev/errorx` errors by passing context which can be used to save caller
+  information
+
 ## v0.1.1 (Released 2024-02-11)
 
 * Fixed bug in signature of JWTAuthService functions